* commit '5c3400da3499b580ae1f9f8b61d1a18fc1b3c72a':
  Do not make provider calls with dummy Accounts.

+++ src/com/android/exchange/adapter/FolderSyncParser.java
@@ -764,6 +764,12 @@ public class FolderSyncParser extends AbstractSyncParser {
 
     @Override
     protected void wipe() {
+        if (mAccountId == EmailContent.NOT_SAVED) {
+            // This is a dummy account so we don't need to do anything yet.
+            return;
+        }
+
+        // For real accounts, let's go ahead and wipe some data.
         EasSyncCalendar.wipeAccountFromContentProvider(mContext,
                 mAccount.mEmailAddress);
         EasSyncContacts.wipeAccountFromContentProvider(mContext,
